requirements.txt
Django==3.2.13
graphene-django==2.15.0

# dev
django-debug-toolbar==3.5.0
django-extensions==3.1.5
